Refactor naming and paths
