+++ tests/Integration/DataFixtures/Command/ServiceChangesCommandTest/test_it_detects_with_inner_substitution_strategy_those_decorated_services_that_changed/git.diff
