--- tests/Stub/ServiceChangesCommand/test_it_detects_directly_decorated_services_that_changed/DecorateOrderEmailManagerInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Webgriffe\SyliusUpgradePlugin\Stub\ServiceChangesCommand\test_it_detects_directly_decorated_services_that_changed;
-
-final class DecorateOrderEmailManagerInterface
-{
-}
