+++ tests/Stub/ServiceChangesCommand/test_it_detects_with_inner_substitution_strategy_those_decorated_services_that_changed/DecorateOrderEmailManagerInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Webgriffe\SyliusUpgradePlugin\Stub\ServiceChangesCommand\test_it_detects_with_inner_substitution_strategy_those_decorated_services_that_changed;
+
+final class DecorateOrderEmailManagerInterface
+{
+}
